+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_101/UpdateMeteringPointConnectionState/StartTrigger_Brs_101_UpdateMeteringPointConnectionState.cs
@@ -14,28 +14,28 @@
 
 using Azure.Messaging.ServiceBus;
 using Energinet.DataHub.Core.Messaging.Communication.Extensions.Options;
-using Energinet.DataHub.ProcessManager.Orchestrations.Extensions.Options;
+using Energinet.DataHub.ProcessManager.Core.Application.Api.Handlers;
+using Energinet.DataHub.ProcessManager.Example.Orchestrations.Extensions.Options;
 using Microsoft.Azure.Functions.Worker;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_026_028.BRS_026.V1;
+namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Processes.BRS_101.UpdateMeteringPointConnectionState;
 
-public class StartTrigger_Brs_026_V1(
-    RequestCalculatedEnergyTimeSeriesHandlerV1 handler)
+internal class StartTrigger_Brs_101_UpdateMeteringPointConnectionState(
+    IStartOrchestrationInstanceFromMessageHandler handler)
 {
-    private readonly RequestCalculatedEnergyTimeSeriesHandlerV1 _handler = handler;
+    private readonly IStartOrchestrationInstanceFromMessageHandler _handler = handler;
 
     /// <summary>
-    /// Start a BRS-026 request.
+    /// Start a BRS-101 Update Metering Point connection state request.
     /// </summary>
-    [Function(nameof(StartTrigger_Brs_026_V1))]
+    [Function(nameof(StartTrigger_Brs_101_UpdateMeteringPointConnectionState))]
     public async Task Run(
         [ServiceBusTrigger(
             $"%{ProcessManagerStartTopicOptions.SectionName}:{nameof(ProcessManagerStartTopicOptions.TopicName)}%",
-            $"%{ProcessManagerStartTopicOptions.SectionName}:{nameof(ProcessManagerStartTopicOptions.Brs026SubscriptionName)}%",
+            $"%{ProcessManagerStartTopicOptions.SectionName}:{nameof(ProcessManagerStartTopicOptions.Brs101UpdateMeteringPointConnectionStateSubscriptionName)}%",
             Connection = ServiceBusNamespaceOptions.SectionName)]
         ServiceBusReceivedMessage message)
     {
-        await _handler.HandleAsync(message)
-            .ConfigureAwait(false);
+        await _handler.HandleAsync(message).ConfigureAwait(false);
     }
 }