+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,17 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
+## Version 2.2.3
+
+- Update `EnqueueMissingMeasurementsLogHttpV1` to contain 1 metering point id for each date.
+
+## Version 2.2.2
+
+- No functional changes.
+
+## Version 2.2.1
+
+- Update `EnqueueMissingMeasurementsLogHttpV1` to contain a list of dates with metering point ids.
+
 ## Version 2.2.0
 
 - Add initial `EnqueueMissingMeasurementsLogHttpV1` for BRS-045.